:recycle: Refactor network entities
